requirements.txt
requests
geopandas
osmnx
networkx
rasterio
matplotlib
